changed more packages to the right format
