+++ docker-compose.yml
@@ -1,9 +1,6 @@
-version: "3"
 services:
   mrobosub:
-    build:
-      context: .
-      dockerfile: Dockerfile
+    image: mrobosub/mrobosub2
     privileged: true
     ports:
       - "10000:10000"
