+++ mrobosub_tests/mrobosub_tests/__init__.py
@@ -1,6 +1,1 @@
-from mrobosub_lib import main
 
-def motor_test():
-    from .motor_test import MotorTest
-
-    main(MotorTest)
